+++ Plugins/BetterImageViewer/CHANGELOG.md
@@ -1,4 +1,9 @@
 # [BetterImageViewer](https://1lighty.github.io/BetterDiscordStuff/?plugin=BetterImageViewer "BetterImageViewer") Changelog
+
+### 1.5.
+- Fixed chat rescale (beta feature) causing embed images to go beyond their borders.
+- Fixed issue with Zere not consistently coding BdApi.getPlugin between BBD and BBD Beta.
+
 ### 1.5.1
 - Fixed chat rescale (beta feature) causing scrolling issues.